requirements.txt
Flask==3.0.0
Pillow==10.0.1
numpy==1.26.0
supabase==2.3.4
tflite-runtime==2.14.0
gunicorn==21.2.0
